+++ src/components/fields/field/index.js
@@ -1,3 +1,4 @@
-﻿import BField from './BField.js';
+import BField from './BField.js';
 
 export { BField, BField as default };
+export { FIELD_KEY, useFieldContext, fieldControlAttrs } from './context.js';