+++ src/components/fields/radio/index.js
@@ -1,0 +1,5 @@
+import BRadio from './BRadio.js';
+import BRadioGroup from './BRadioGroup.js';
+
+export { BRadio, BRadioGroup, BRadio as default };
+export { RADIO_GROUP_KEY } from './context.js';