+++ src/components/fields/select/index.js
@@ -1,0 +1,3 @@
+import BSelect from './BSelect.js';
+
+export { BSelect, BSelect as default };