+++ src/components/fields/switch/index.js
@@ -1,0 +1,3 @@
+import BSwitch from './BSwitch.js';
+
+export { BSwitch, BSwitch as default };