+++ src/components/fields/textarea/index.js
@@ -1,0 +1,3 @@
+import BTextarea from './BTextarea.js';
+
+export { BTextarea, BTextarea as default };