+++ src/vue.js
@@ -1,4 +1,4 @@
-﻿import { BGridLayout, BGridArea } from './components/containers/grid/index.js';
+import { BGridLayout, BGridArea } from './components/containers/grid/index.js';
 import { BWrapper } from './components/containers/wrapper/index.js';
 import { BButton } from './components/cta/button/index.js';
 import { BClickableArea } from './components/cta/clickable-area/index.js';
@@ -7,7 +7,11 @@ import { BMenu } from './components/nav/menu/index.js';
 import { BPagination } from './components/nav/pagination/index.js';
 import { BField } from './components/fields/field/index.js';
 import { BInput } from './components/fields/input/index.js';
+import { BSelect } from './components/fields/select/index.js';
+import { BTextarea } from './components/fields/textarea/index.js';
 import { BCheckbox } from './components/fields/checkbox/index.js';
+import { BRadio, BRadioGroup } from './components/fields/radio/index.js';
+import { BSwitch } from './components/fields/switch/index.js';
 import { BBadge } from './components/info/badge/index.js';
 import { BImage } from './components/media/image/index.js';
 import { BLogo } from './components/brand/logo/index.js';
@@ -22,7 +26,12 @@ export {
   BMenu,
   BField,
   BInput,
+  BSelect,
+  BTextarea,
   BCheckbox,
+  BRadio,
+  BRadioGroup,
+  BSwitch,
   BPagination,
   BBadge,
   BImage,